// src/api/zzirit/models/BrandFetchResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * zzirit API
 * 찌릿 API 명세서
 *
 * The version of the OpenAPI document: v0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface BrandFetchResponse
 */
export interface BrandFetchResponse {
    /**
     * 
     * @type {number}
     * @memberof BrandFetchResponse
     */
    brandId?: number;
    /**
     * 
     * @type {string}
     * @memberof BrandFetchResponse
     */
    name?: string;
}

/**
 * Check if a given object implements the BrandFetchResponse interface.
 */
export function instanceOfBrandFetchResponse(value: object): value is BrandFetchResponse {
    return true;
}

export function BrandFetchResponseFromJSON(json: any): BrandFetchResponse {
    return BrandFetchResponseFromJSONTyped(json, false);
}

export function BrandFetchResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): BrandFetchResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'brandId': json['brandId'] == null ? undefined : json['brandId'],
        'name': json['name'] == null ? undefined : json['name'],
    };
}

export function BrandFetchResponseToJSON(json: any): BrandFetchResponse {
    return BrandFetchResponseToJSONTyped(json, false);
}

export function BrandFetchResponseToJSONTyped(value?: BrandFetchResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'brandId': value['brandId'],
        'name': value['name'],
    };
}